--- apps/src/codebridge/InfoPanel/styles/for-teachers-only.module.scss
@@ -1,12 +1,0 @@
-@import 'color.scss';
-
-.predictSolutionContainer {
-  margin: 10px;
-  padding: 5px;
-  background-color: $light_white;
-  border-radius: 5px;
-}
-
-.multipleChoiceLetter {
-  font-weight: bold;
-}
