+++ apps/src/lab2/views/components/PredictQuestion.tsx
@@ -3,7 +3,7 @@ import {
   LevelPredictSettings,
   PredictQuestionType,
 } from '@cdo/apps/lab2/levelEditors/types';
-import moduleStyles from './predict-question.module.scss';
+import moduleStyles from './predict.module.scss';
 
 interface PredictQuestionProps {
   predictSettings: LevelPredictSettings | undefined;
